test(material/progress-spinner): combine shared tests
Since we only have one module, we don't need separate shared tests anymore.
